src/scrapers/company_scraper.py
"""
公司信息爬虫模块
"""
import asyncio
import aiohttp
from bs4 import BeautifulSoup
from typing import Optional
from loguru import logger
from config import settings


class CompanyScraper:
    """公司信息爬虫"""
    
    def __init__(self):
        self.timeout = aiohttp.ClientTimeout(total=settings.CRAWLER_TIMEOUT)
        self.headers = {
            "User-Agent": settings.CRAWLER_USER_AGENT
        }
    
    async def scrape(
        self,
        company_name: str,
        url: Optional[str] = None,
        include_recruitment: bool = True
    ) -> dict:
        """
        爬取公司信息
        
        Args:
            company_name: 公司名称
            url: 公司官网URL（可选）
            include_recruitment: 是否包含招聘信息
        
        Returns:
            公司信息字典
        """
        try:
            result = {
                "company_name": company_name,
                "url": url,
                "basic_info": {},
                "culture": {},
                "positions": []
            }
            
            # 如果没有提供URL，尝试搜索
            if not url:
                url = await self._search_company_url(company_name)
                result["url"] = url
            
            if url:
                # 爬取基本信息
                result["basic_info"] = await self._scrape_basic_info(url)
                
                # 爬取企业文化
                result["culture"] = await self._scrape_culture(url)
                
                # 爬取招聘信息
                if include_recruitment:
                    result["positions"] = await self._scrape_positions(url, company_name)
            
            logger.info(f"成功爬取公司信息: {company_name}")
            return result
            
        except Exception as e:
            logger.error(f"爬取公司信息失败: {company_name}, 错误: {e}")
            return {
                "company_name": company_name,
                "error": str(e),
                "basic_info": {},
                "culture": {},
                "positions": []
            }
    
    async def _search_company_url(self, company_name: str) -> Optional[str]:
        """搜索公司官网URL"""
        # TODO: 实现搜索逻辑（可以调用搜索引擎API或使用已知的公司URL映射）
        logger.info(f"搜索公司URL: {company_name}")
        
        # 示例：常见公司URL映射
        known_companies = {
            "腾讯": "https://www.tencent.com",
            "阿里巴巴": "https://www.alibaba.com",
            "字节跳动": "https://www.bytedance.com",
            "百度": "https://www.baidu.com",
            "华为": "https://www.huawei.com",
        }
        
        return known_companies.get(company_name)
    
    async def _scrape_basic_info(self, url: str) -> dict:
        """爬取公司基本信息"""
        try:
            async with aiohttp.ClientSession(timeout=self.timeout) as session:
                async with session.get(url, headers=self.headers) as response:
                    if response.status == 200:
                        html = await response.text()
                        soup = BeautifulSoup(html, 'lxml')
                        
                        # 提取基本信息（需要根据实际网站结构调整）
                        info = {
                            "description": self._extract_description(soup),
                            "industry": self._extract_industry(soup),
                            "size": self._extract_size(soup),
                            "founded": self._extract_founded(soup)
                        }
                        
                        return info
        except Exception as e:
            logger.error(f"爬取基本信息失败: {url}, 错误: {e}")
        
        return {}
    
    async def _scrape_culture(self, url: str) -> dict:
        """爬取企业文化"""
        try:
            # TODO: 实现企业文化爬取逻辑
            return {
                "values": [],
                "mission": "",
                "vision": ""
            }
        except Exception as e:
            logger.error(f"爬取企业文化失败: {url}, 错误: {e}")
            return {}
    
    async def _scrape_positions(self, url: str, company_name: str) -> list:
        """爬取招聘信息"""
        try:
            # TODO: 实现招聘信息爬取
            # 可以爬取公司官网的招聘页面，或者调用招聘平台API
            
            # 示例：返回模拟数据
            positions = [
                {
                    "title": "Python后端工程师",
                    "location": "北京",
                    "salary": "20-35K",
                    "requirements": [
                        "3年以上Python开发经验",
                        "熟悉Django/Flask框架",
                        "熟悉MySQL、Redis等数据库"
                    ],
                    "responsibilities": [
                        "负责后端服务开发和维护",
                        "参与系统架构设计"
                    ]
                }
            ]
            
            return positions
            
        except Exception as e:
            logger.error(f"爬取招聘信息失败: {url}, 错误: {e}")
            return []
    
    def _extract_description(self, soup: BeautifulSoup) -> str:
        """提取公司描述"""
        # TODO: 根据实际网站结构提取
        return ""
    
    def _extract_industry(self, soup: BeautifulSoup) -> str:
        """提取行业信息"""
        return ""
    
    def _extract_size(self, soup: BeautifulSoup) -> str:
        """提取公司规模"""
        return ""
    
    def _extract_founded(self, soup: BeautifulSoup) -> str:
        """提取成立时间"""
        return ""
