requirements.txt
# MCP协议相关
mcp>=0.9.0

# 网络爬虫
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
scrapy>=2.11.0
lxml>=4.9.0

# AI模型
ollama>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10

# 简历解析
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0
pytesseract>=0.3.10

# 数据处理
pandas>=2.1.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
aiohttp>=3.9.0

# 日志和监控
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
